♻️ refactor: pricesection 컴포넌트 분리
